requirements.txt
aws-cdk-lib==2.147.0
constructs>=10.0.0,<11.0.0
boto3==1.34.130
boto3[crt] ==1.16.0
